# api/docs/swagger.yaml
basePath: /api/v1
info:
  contact:
    email: pejman.hkh@gmail.com
    name: Pejman Hkh
    url: https://www.peji.ir
  description: Subtitle api
  title: Subtitle API
  version: "1.0"
paths:
  /movie/{link}:
    get:
      consumes:
      - application/json
      description: Get movie details
      parameters:
      - description: Link
        in: path
        name: link
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Details movie
      tags:
      - movies
  /movie/{link}/{season}:
    get:
      consumes:
      - application/json
      description: Get Season
      parameters:
      - description: Link
        in: path
        name: link
        required: true
        type: string
      - description: Season
        in: path
        name: season
        required: true
        type: number
      produces:
      - application/json
      responses: {}
      summary: Get Season
      tags:
      - movies
  /movies:
    get:
      consumes:
      - application/json
      description: Get list of all movies
      parameters:
      - description: Search
        in: query
        name: search
        type: string
      - description: Pagination
        in: query
        name: page
        type: number
      - description: Name
        in: query
        name: name
        type: string
      - description: Year
        in: query
        name: year
        type: number
      - description: Imdb Code
        in: query
        name: imdb_code
        type: string
      - description: Type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses: {}
      summary: List movies
      tags:
      - movies
  /movies/detail:
    get:
      consumes:
      - application/json
      description: Get movie details
      parameters:
      - description: ID
        in: query
        name: id
        type: integer
      - description: IMDB Code
        in: query
        name: imdb
        type: string
      produces:
      - application/json
      responses: {}
      summary: Details movie
      tags:
      - movies
  /movies/search:
    get:
      consumes:
      - application/json
      description: Search in movies
      parameters:
      - description: Search
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses: {}
      summary: Search movies
      tags:
      - movies
  /subtitles/{id}/download:
    get:
      consumes:
      - application/json
      description: Download Subtitle
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Download Subtitle
      tags:
      - Subtitles
  /subtitles/{id}/json:
    get:
      consumes:
      - application/json
      description: Json Subtitle
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Json Subtitle
      tags:
      - Subtitles
swagger: "2.0"